+++ packages/commons/CHANGELOG.md
@@ -1,3 +1,9 @@
+## [0.0.128](https://github.com/doomsower/whitewater/compare/@whitewater-guide/commons@0.0.127...@whitewater-guide/commons@0.0.128) (2019-04-26)
+
+### Bug Fixes
+
+- **commons:** ignore 0 rating when filtering ([4ca3370](https://github.com/doomsower/whitewater/commit/4ca3370))
+
 ## [0.0.127](https://github.com/doomsower/whitewater/compare/@whitewater-guide/commons@0.0.126...@whitewater-guide/commons@0.0.127) (2019-04-12)
 
 ## [0.0.126](https://github.com/doomsower/whitewater/compare/@whitewater-guide/commons@0.0.125...@whitewater-guide/commons@0.0.126) (2019-04-09)